+++ appstream-center-20210901/alibabacloud_appstream_center20210901/models/_list_wuying_server_response_body.py
@@ -83,11 +83,14 @@ class ListWuyingServerResponseBodyWuyingServerList(DaraModel):
     def __init__(
         self,
         add_virtual_node_pool_status: str = None,
+        ali_uid: int = None,
+        bandwidth: int = None,
         biz_region_id: str = None,
         charge_type: str = None,
         create_time: str = None,
         data_disk: List[main_models.ListWuyingServerResponseBodyWuyingServerListDataDisk] = None,
         expired_time: str = None,
+        fota_version: str = None,
         image_id: str = None,
         image_name: str = None,
         instance_info_list: List[main_models.ListWuyingServerResponseBodyWuyingServerListInstanceInfoList] = None,
@@ -97,19 +100,27 @@ class ListWuyingServerResponseBodyWuyingServerList(DaraModel):
         office_site_name: str = None,
         office_site_type: str = None,
         os_type: str = None,
+        policy_group_id_list: List[str] = None,
+        resource_session_status: str = None,
         security_group_ids: List[str] = None,
         server_instance_type_info: main_models.ListWuyingServerResponseBodyWuyingServerListServerInstanceTypeInfo = None,
+        sessions: List[main_models.ListWuyingServerResponseBodyWuyingServerListSessions] = None,
         status: str = None,
         sub_pay_type: str = None,
         system_disk_category: str = None,
+        system_disk_id: str = None,
         system_disk_performance_level: str = None,
         system_disk_size: int = None,
+        timer_group_id: str = None,
+        users: List[str] = None,
         virtual_kubelet_ip: str = None,
         virtual_node_pool_id: str = None,
         wuying_server_id: str = None,
         wuying_server_name: str = None,
     ):
         self.add_virtual_node_pool_status = add_virtual_node_pool_status
+        self.ali_uid = ali_uid
+        self.bandwidth = bandwidth
         # Region.
         self.biz_region_id = biz_region_id
         # The billing method.
@@ -120,6 +131,7 @@ class ListWuyingServerResponseBodyWuyingServerList(DaraModel):
         self.data_disk = data_disk
         # The time when the subscription instance expires.
         self.expired_time = expired_time
+        self.fota_version = fota_version
         # The ID of the custom image.
         self.image_id = image_id
         # The image name.
@@ -137,18 +149,24 @@ class ListWuyingServerResponseBodyWuyingServerList(DaraModel):
         self.office_site_type = office_site_type
         # The OS type.
         self.os_type = os_type
+        self.policy_group_id_list = policy_group_id_list
+        self.resource_session_status = resource_session_status
         self.security_group_ids = security_group_ids
         # The specifications.
         self.server_instance_type_info = server_instance_type_info
+        self.sessions = sessions
         # The status of the workstation.
         self.status = status
         self.sub_pay_type = sub_pay_type
         # The type of the system disk.
         self.system_disk_category = system_disk_category
+        self.system_disk_id = system_disk_id
         # The performance level (PL) of the system disk.
         self.system_disk_performance_level = system_disk_performance_level
         # The size of the system disk. Unit: GiB.
         self.system_disk_size = system_disk_size
+        self.timer_group_id = timer_group_id
+        self.users = users
         self.virtual_kubelet_ip = virtual_kubelet_ip
         self.virtual_node_pool_id = virtual_node_pool_id
         # The ID of the workstation.
@@ -167,6 +185,10 @@ class ListWuyingServerResponseBodyWuyingServerList(DaraModel):
                     v1.validate()
         if self.server_instance_type_info:
             self.server_instance_type_info.validate()
+        if self.sessions:
+            for v1 in self.sessions:
+                 if v1:
+                    v1.validate()
 
     def to_map(self):
         result = dict()
@@ -175,6 +197,12 @@ class ListWuyingServerResponseBodyWuyingServerList(DaraModel):
             result = _map
         if self.add_virtual_node_pool_status is not None:
             result['AddVirtualNodePoolStatus'] = self.add_virtual_node_pool_status
+
+        if self.ali_uid is not None:
+            result['AliUid'] = self.ali_uid
+
+        if self.bandwidth is not None:
+            result['Bandwidth'] = self.bandwidth
 
         if self.biz_region_id is not None:
             result['BizRegionId'] = self.biz_region_id
@@ -192,6 +220,9 @@ class ListWuyingServerResponseBodyWuyingServerList(DaraModel):
 
         if self.expired_time is not None:
             result['ExpiredTime'] = self.expired_time
+
+        if self.fota_version is not None:
+            result['FotaVersion'] = self.fota_version
 
         if self.image_id is not None:
             result['ImageId'] = self.image_id
@@ -222,11 +253,22 @@ class ListWuyingServerResponseBodyWuyingServerList(DaraModel):
         if self.os_type is not None:
             result['OsType'] = self.os_type
 
+        if self.policy_group_id_list is not None:
+            result['PolicyGroupIdList'] = self.policy_group_id_list
+
+        if self.resource_session_status is not None:
+            result['ResourceSessionStatus'] = self.resource_session_status
+
         if self.security_group_ids is not None:
             result['SecurityGroupIds'] = self.security_group_ids
 
         if self.server_instance_type_info is not None:
             result['ServerInstanceTypeInfo'] = self.server_instance_type_info.to_map()
+
+        result['Sessions'] = []
+        if self.sessions is not None:
+            for k1 in self.sessions:
+                result['Sessions'].append(k1.to_map() if k1 else None)
 
         if self.status is not None:
             result['Status'] = self.status
@@ -237,11 +279,20 @@ class ListWuyingServerResponseBodyWuyingServerList(DaraModel):
         if self.system_disk_category is not None:
             result['SystemDiskCategory'] = self.system_disk_category
 
+        if self.system_disk_id is not None:
+            result['SystemDiskId'] = self.system_disk_id
+
         if self.system_disk_performance_level is not None:
             result['SystemDiskPerformanceLevel'] = self.system_disk_performance_level
 
         if self.system_disk_size is not None:
             result['SystemDiskSize'] = self.system_disk_size
+
+        if self.timer_group_id is not None:
+            result['TimerGroupId'] = self.timer_group_id
+
+        if self.users is not None:
+            result['Users'] = self.users
 
         if self.virtual_kubelet_ip is not None:
             result['VirtualKubeletIp'] = self.virtual_kubelet_ip
@@ -262,6 +313,12 @@ class ListWuyingServerResponseBodyWuyingServerList(DaraModel):
         if m.get('AddVirtualNodePoolStatus') is not None:
             self.add_virtual_node_pool_status = m.get('AddVirtualNodePoolStatus')
 
+        if m.get('AliUid') is not None:
+            self.ali_uid = m.get('AliUid')
+
+        if m.get('Bandwidth') is not None:
+            self.bandwidth = m.get('Bandwidth')
+
         if m.get('BizRegionId') is not None:
             self.biz_region_id = m.get('BizRegionId')
 
@@ -279,6 +336,9 @@ class ListWuyingServerResponseBodyWuyingServerList(DaraModel):
 
         if m.get('ExpiredTime') is not None:
             self.expired_time = m.get('ExpiredTime')
+
+        if m.get('FotaVersion') is not None:
+            self.fota_version = m.get('FotaVersion')
 
         if m.get('ImageId') is not None:
             self.image_id = m.get('ImageId')
@@ -310,12 +370,24 @@ class ListWuyingServerResponseBodyWuyingServerList(DaraModel):
         if m.get('OsType') is not None:
             self.os_type = m.get('OsType')
 
+        if m.get('PolicyGroupIdList') is not None:
+            self.policy_group_id_list = m.get('PolicyGroupIdList')
+
+        if m.get('ResourceSessionStatus') is not None:
+            self.resource_session_status = m.get('ResourceSessionStatus')
+
         if m.get('SecurityGroupIds') is not None:
             self.security_group_ids = m.get('SecurityGroupIds')
 
         if m.get('ServerInstanceTypeInfo') is not None:
             temp_model = main_models.ListWuyingServerResponseBodyWuyingServerListServerInstanceTypeInfo()
             self.server_instance_type_info = temp_model.from_map(m.get('ServerInstanceTypeInfo'))
+
+        self.sessions = []
+        if m.get('Sessions') is not None:
+            for k1 in m.get('Sessions'):
+                temp_model = main_models.ListWuyingServerResponseBodyWuyingServerListSessions()
+                self.sessions.append(temp_model.from_map(k1))
 
         if m.get('Status') is not None:
             self.status = m.get('Status')
@@ -326,11 +398,20 @@ class ListWuyingServerResponseBodyWuyingServerList(DaraModel):
         if m.get('SystemDiskCategory') is not None:
             self.system_disk_category = m.get('SystemDiskCategory')
 
+        if m.get('SystemDiskId') is not None:
+            self.system_disk_id = m.get('SystemDiskId')
+
         if m.get('SystemDiskPerformanceLevel') is not None:
             self.system_disk_performance_level = m.get('SystemDiskPerformanceLevel')
 
         if m.get('SystemDiskSize') is not None:
             self.system_disk_size = m.get('SystemDiskSize')
+
+        if m.get('TimerGroupId') is not None:
+            self.timer_group_id = m.get('TimerGroupId')
+
+        if m.get('Users') is not None:
+            self.users = m.get('Users')
 
         if m.get('VirtualKubeletIp') is not None:
             self.virtual_kubelet_ip = m.get('VirtualKubeletIp')
@@ -346,12 +427,48 @@ class ListWuyingServerResponseBodyWuyingServerList(DaraModel):
 
         return self
 
+class ListWuyingServerResponseBodyWuyingServerListSessions(DaraModel):
+    def __init__(
+        self,
+        resource_session_start_time: str = None,
+        user_id: str = None,
+    ):
+        self.resource_session_start_time = resource_session_start_time
+        self.user_id = user_id
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        result = dict()
+        _map = super().to_map()
+        if _map is not None:
+            result = _map
+        if self.resource_session_start_time is not None:
+            result['ResourceSessionStartTime'] = self.resource_session_start_time
+
+        if self.user_id is not None:
+            result['UserId'] = self.user_id
+
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('ResourceSessionStartTime') is not None:
+            self.resource_session_start_time = m.get('ResourceSessionStartTime')
+
+        if m.get('UserId') is not None:
+            self.user_id = m.get('UserId')
+
+        return self
+
 class ListWuyingServerResponseBodyWuyingServerListServerInstanceTypeInfo(DaraModel):
     def __init__(
         self,
         cpu: str = None,
         gpu: str = None,
         gpu_memory: int = None,
+        gpu_spec: str = None,
         memory: int = None,
         server_instance_type: str = None,
     ):
@@ -361,6 +478,7 @@ class ListWuyingServerResponseBodyWuyingServerListServerInstanceTypeInfo(DaraMod
         self.gpu = gpu
         # The memory size. Unit: MB.
         self.gpu_memory = gpu_memory
+        self.gpu_spec = gpu_spec
         # The memory size. Unit: MB.
         self.memory = memory
         # Workstation specifications.
@@ -383,6 +501,9 @@ class ListWuyingServerResponseBodyWuyingServerListServerInstanceTypeInfo(DaraMod
         if self.gpu_memory is not None:
             result['GpuMemory'] = self.gpu_memory
 
+        if self.gpu_spec is not None:
+            result['GpuSpec'] = self.gpu_spec
+
         if self.memory is not None:
             result['Memory'] = self.memory
 
@@ -401,6 +522,9 @@ class ListWuyingServerResponseBodyWuyingServerListServerInstanceTypeInfo(DaraMod
 
         if m.get('GpuMemory') is not None:
             self.gpu_memory = m.get('GpuMemory')
+
+        if m.get('GpuSpec') is not None:
+            self.gpu_spec = m.get('GpuSpec')
 
         if m.get('Memory') is not None:
             self.memory = m.get('Memory')
@@ -451,11 +575,15 @@ class ListWuyingServerResponseBodyWuyingServerListDataDisk(DaraModel):
     def __init__(
         self,
         data_disk_category: str = None,
+        data_disk_id: str = None,
+        data_disk_no: str = None,
         data_disk_performance_level: str = None,
         data_disk_size: int = None,
     ):
         # The category of data disk.
         self.data_disk_category = data_disk_category
+        self.data_disk_id = data_disk_id
+        self.data_disk_no = data_disk_no
         # The PL of the data disk.
         self.data_disk_performance_level = data_disk_performance_level
         # The size of the data disk. Unit: GB.
@@ -472,6 +600,12 @@ class ListWuyingServerResponseBodyWuyingServerListDataDisk(DaraModel):
         if self.data_disk_category is not None:
             result['DataDiskCategory'] = self.data_disk_category
 
+        if self.data_disk_id is not None:
+            result['DataDiskId'] = self.data_disk_id
+
+        if self.data_disk_no is not None:
+            result['DataDiskNo'] = self.data_disk_no
+
         if self.data_disk_performance_level is not None:
             result['DataDiskPerformanceLevel'] = self.data_disk_performance_level
 
@@ -484,6 +618,12 @@ class ListWuyingServerResponseBodyWuyingServerListDataDisk(DaraModel):
         m = m or dict()
         if m.get('DataDiskCategory') is not None:
             self.data_disk_category = m.get('DataDiskCategory')
+
+        if m.get('DataDiskId') is not None:
+            self.data_disk_id = m.get('DataDiskId')
+
+        if m.get('DataDiskNo') is not None:
+            self.data_disk_no = m.get('DataDiskNo')
 
         if m.get('DataDiskPerformanceLevel') is not None:
             self.data_disk_performance_level = m.get('DataDiskPerformanceLevel')